build: Do not use the linting-tsconfig for emitting files
Only use "tsconfig.build.json" for building and emitting. Use
"tsconfig.json" for checking for compiler errors and for integrating
with typescript-eslint.

The directory structure emitted when using "tsconfig.json" is different
from the regular build in the sense that it contains an additional "src"
directory level.

Configuring "noEmit: true" in the tsconfig file causes issues for the
typescript-eslint integration. Therefore the flag is only set in the npm
script.