ci: Skip test-build in test matrix
The "test-build" script compiles the code to check for TypeScript
compilation errors. This is not specific to Node.js / OS and can
therefore be skipped in the matrix. It is also no prerequisite for
running the unit tests, as ava compiles them on-the-fly.
The "test-build" script is still part of the ci.yml workflow.