refactor: Enable eslint RegEx rules (#240)
With the introduction of ESLint 9 and the update of `typescript-eslint`,
there are new rules that will be applied to the code. Those rules are
mainly focused on Regex patterns.

With https://github.com/SAP/ui5-linter/pull/230 those rules have been
turned off to keep it consistent with the old behavior.
This change aims to enable those rules.

**Note:** The base for this change is
https://github.com/SAP/ui5-linter/pull/230 and not the **main** branch.
This way it's clearer what the real changes are.