feat: Detect partially deprecated APIs (#286)
Detect a set of prominent partially-deprecated API (e.g. deprecation of a single parameter of a method signature):

* sap/ui/core/theming/Parameters.get
* sap/ui/core/Component.createComponent (async flag is currently not described as deprecated, but is likely to
be done soon)
* sap/ui/core/routing/Router (async flag)
* sap/ui/model/json/JSONModel#loadData
* sap/ui/model/odata/v4/ODataModel constructor
* sap/ui/model/odata/v2/ODataModel#createEntry
* sap/ui/util/Mobile.init

Generic detection would be better, ideally expressed through the
generation of the respective TypeScript type definitions. However, that
would first require changes in the UI5 JSDoc plugin in order to allow
developers to properly express deprecated parameters. Discussions on
this are ongoing. For the time being this is a simple solution to detect
widely used deprecated parameters.

JIRA: CPOUI5FOUNDATION-788

---------

Co-authored-by: Matthias Osswald <mat.osswald@sap.com>