test: Add fixture for sap.ui.suite project
There are some special circumstances when linting sap.ui.suite which
were detected by manual testing.
Some findings were only reported with a newer version of UI5 linter.

It turned out that https://github.com/SAP/ui5-linter/pull/290 changed
the behavior and fixed the reporting of deprecations within the same
library, as previously the @sapui5/types were not loaded properly.

It is still unknown why exactly this was happening, but it is fixed now
and the added test project should ensure that no similar issues will
appear in the future.