feat: Improve code detection for UI5 classes (JavaScript)
This change improves the detection of JavaScript code related to UI5
classes written in the scanned project by converting all UI5 classes to
ES classes.

Previously, the linter only converted some of the UI5 classes which left
some code uncovered as the linter did not recognize the proper types.

This is also a preparation to enable further checks on sub-classes of
UI5 framework base classes such as controls and elements.