fix(amdTranspiler): Fix exception "Node not found in array"
When multiple return statements are found in the AMD factory, the
factory function will not be replaced, so no assignment of the old
factory block should be done.
This caused the exception "Node not found in array" as replacements
were wrongly applied on the new top level SourceFile instead of the
original and still existing factory function.