fix: Properly check UI5 class inheritance
The UI5 class inheritance check needs to perform a different check when
the UI5 base classes are actually within the current project like it is
the case for the sap.ui.core library.
In this case the declaration is based on the actual code, not the UI5
type definitions.

This change adds a very minimal sap.ui.core project where this scenario
can be reproduced without having to test the whole original library.

This is a follow-up of https://github.com/SAP/ui5-linter/pull/349