refactor: Update api-extract logic for deprecations
This is the starting point of a larger refactoring to optimize the
structure of the api-extract.json file and the consumption of the
included info in the linter code.

This change focuses on extracting the same information during the
existing iteration over all symbols so that information such as
default aggregation, deprecation text and more can be stored at the
symbol rather than having to repeat symbol names for every kind of
information that needs to be accessed.

It also adds missing deprecation texts for some cases.