fix: Properly check directly exported renderer
There was a gap with the detection of Control renderers that are
declared in a separate module and directly returned within the module
without using a locale variable.
This issue was only reproducible within the context of UI5 framework
libraries where type definitions via @sapui5/types are available in
addition to the actual source code.
For that reason a specific test case within the sap.f test project has
been added.