fix: False-positive global findings in controllers
This solves an issue where the linter would falsely report access to
global variables in controller code when accessing the controller class
(e.g. by adding methods to the prototype).

The issue was caused by the introduction of type support for "byId" in
controllers and only appears when there is a view that references the
controller.

Caused by: https://github.com/SAP/ui5-linter/pull/423