fix: Improve module resolution
This solves issues where some detections did not work properly due to
the way dependencies were resolved.

When using relative imports, the module resolution was working, but when
for example importing a control from a test, the complete namespace is
used, but the corresponding JS file was not found.

This change ensures that JS files are resolved and also that the types
for a framework library are loaded when it is linted.