fix(ConfigManager): Resolve ESM import exception on Windows
import(path) does not accept absolute Windows-paths since those seem to
get confused with URLs.

Before, this was mitigated by always calculating a relative path
(relative to the UI5 linter install directory). This however breaks on
Windows if the target project is located on another drive as reported in
https://github.com/SAP/ui5-linter/issues/458

Instead convert all paths to "file://"-URLs before passing them to
import().

Also, in linter.ts, enforce the project root path to always be absolute
to simplify handling the path later on.

Resolves https://github.com/SAP/ui5-linter/issues/458