feat: Improve deprecation detection of property accesses
When a deprecated API is accessed as part of the arguments of a function
call, the linter did not detect it.
With this change, all property accesses of deprecated APIs are detected.

In addition, the namespace/usage of the deprecated property is included
in the message which gives more context about the finding without
looking into the code.