fix: Improve overall code detection (#489)
This change improves the overall code detection by refactoring the
SourceFileLinter class to make use of utility functions that increase
the coverage of AST node types that may appear and have not been checked
so far. This mainly includes TemplateLiterals by using
`isStringLiteralLike` instead of `isStringLiteral`.

In addition, the usage of `getText` has been replaced by actual node
checks and access to the actual properties (e.g. `text`). This makes
handling of quotes obsolete and is safer as whitespaces or line breaks
within the code might be included in the result of `getText`.