fix: Improve deprecation detection
This change improves the detection of deprecated API usage so that more
scenarios are covered.

Previously it may have happened that the wrong symbol was used to check
for JSDoc comments in case a value is re-exported and only the re-export
is marked as deprecated.

This change also improves the detection of named imports in TypeScript
and removes checks for cases that do not appear to be relevant.