fix(amdTranspiler): Ensure correct line information
In some cases the line information of linter findings was not correct.
The information was only updated as a side-effect of calling `toPosStr`
when throwing an UnsupportedExtendCall exception.

With this change, the line information is always updated when
transforming a source file by calling "getLineStarts".

This change also prevents excessive exception logging for UI5 extend
calls in case no actual "extend" call is visited, which should also
improve overall performance.