fix: Cannot read properties of undefined (reading 'map')
This fixes an exception that was happening when a source map with
sections was being processed.
It mainly occurred when a UI5 bundle such as a `Component-preload.js` is
part of the project sources and not excluded from linting.

UI5 bundles are now excluded from linting and AMD transpilation.
In addition, source maps with sections are not processed.

Fixes: https://github.com/SAP/ui5-linter/issues/583