perf: Improve lazy loading of TypeScript definitions
With the introduction of types for pseudo modules, the lazy loading of
framework library types did not work as expected anymore as all pseudo
module types did eagerly load the library types.

This change introduces a new entrypoint for each library, which can be
lazy loaded via paths mappings.
The entrypoint takes care of loading the actual types of a library such
as the @sapui5/types file and pseudo modules, if they exist.

In addition, dependencies from a manifest.json file are now also
considered when loading types.

When using the UI5LinterEngine, the common types are not released
anymore when they are not used be the subsequent program.
This improves the reuse of common types which improves performance when
linting multiple projects via the same engine instance.