fix(autofix): Do not fix global access of Core module (#641)
The global access provides the Core class, while the module import
only provides the singleton instance.

While some cases would be fixable, the usage of the Core class as global
was never a best practices and is seen as very uncommon. For that reason
no fix is provided and the message will remain to be solved manually.