refactor: Introduce Ui5TypeInfo
Introduces a Ui5TypeInfo which might be especially useful in future for
usage in the autofix context.

The info will be collected when checking for deprecations and attached
to the raw messages. For now, the information is only available
internally but exposing it via the linter messages might be done later
to allow consumers to use it.