refactor(autofix): Allow fixes to keep current property access
By not specifying a moduleName or globalName, it is assumed that the fix
requests to keep the current property access. E.g. "module" in
"module.method()".

This is required in cases where an API was renamed on the same
module/class and the property access needs to stay the same since it
might be a variable holding an instance of a class.